api/requirements.txt
pillow~=10.1.0
fastapi~=0.104.1
pydantic~=2.4.2
uvicorn
python-multipart
numpy~=1.26.1
scikit-image~=0.22.0
opencv-python~=4.8.1.78
SQLAlchemy~=2.0.23
jinja2
